% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping.r
\name{argErrorExternalObjType}
\alias{argErrorExternalObjType}
\title{Build an error message that an unexpected object is in lieu of an expected cinterop external ref object.}
\usage{
argErrorExternalObjType(x, expectedType)
}
\arguments{
\item{x}{actual object that is not of the expected type or underlying type for the external pointer.}

\item{expectedType}{expected underlying type for the ExternalObj}
}
\value{
a character, the error message
}
\description{
Build an error message that an unexpected object is in lieu of an expected cinterop external ref object.
}
