% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.r
\name{asInteropRegularTimeSeries}
\alias{asInteropRegularTimeSeries}
\title{Convert an xts object to a cinterop RegularTimeSeries object to be passed to a C API}
\usage{
asInteropRegularTimeSeries(xts_mts)
}
\arguments{
\item{xts_mts}{the uni- or multi-variate xts time series}
}
\value{
an object of class "RegularTimeSeries"
}
\description{
Convert an xts object to a cinterop RegularTimeSeries object to be passed to a C API
}

