% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00class-definitions.r
\docType{class}
\name{ExternalObjRef-class}
\alias{ExternalObjRef-class}
\title{Class ExternalObjRef.}
\description{
Class \code{ExternalObjRef} defines a class that is used to hold references to external objects accessed via opaque, external pointers.
}
