% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping.r
\name{getExternalXptr}
\alias{getExternalXptr}
\title{Gets an externalptr wrapped in an ExternalObjRef}
\usage{
getExternalXptr(objRef, stringent = FALSE)
}
\arguments{
\item{objRef}{the presumed ExternalObjRef to unwrap}

\item{stringent}{if TRUE, an error is raised if objRef is neither an  ExternalObjRef nor an externalptr.}
}
\value{
an externalptr, or the input objRef unchanged if objRef is neither an  ExternalObjRef nor an externalptr and not in stringent mode
}
\description{
Gets an externalptr wrapped in an ExternalObjRef
}
