% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping.r
\name{isExternalObjRef}
\alias{isExternalObjRef}
\title{Is an object a cinterop wrapper}
\usage{
isExternalObjRef(x, type)
}
\arguments{
\item{x}{The object to check: is it an ExternalObjRef}

\item{type}{optional, the exact type of the ExternalObjRef to expect}
}
\value{
a logical value
}
\description{
Is an object a cinterop wrapper, and if so of a specified type
}
