% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.r
\name{isInteropRegularTimeSeries}
\alias{isInteropRegularTimeSeries}
\title{Is the object a cinterop RegularTimeSeries object to be passed to a C API}
\usage{
isInteropRegularTimeSeries(tsInfo)
}
\arguments{
\item{tsInfo}{the object to test}
}
\description{
Is the object a cinterop RegularTimeSeries object to be passed to a C API
}

