% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.r
\name{makeMonthlyTimeAxis}
\alias{makeMonthlyTimeAxis}
\title{Creates a vector that can be used as a monthly time series index for an xts series}
\usage{
makeMonthlyTimeAxis(startDate, n)
}
\arguments{
\item{startDate}{a date object such as a POSIXct}

\item{n}{number of months}
}
\description{
Creates a vector that can be used as a monthly time series index for an xts series
}
