% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.r
\name{makeTimeAxis}
\alias{makeTimeAxis}
\title{Creates an R posixCT time axis for use in e.g. xts indexing}
\usage{
makeTimeAxis(tsGeom)
}
\arguments{
\item{tsGeom}{an S4 object of class 'RegularTimeSeriesGeometry'}
}
\description{
Creates an R posixCT time axis for use in e.g. xts indexing
}
