% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping.r
\name{mkExternalObjRef}
\alias{mkExternalObjRef}
\title{Create a wrapper around an externalptr}
\usage{
mkExternalObjRef(obj, type = "")
}
\arguments{
\item{obj}{the object to wrap, if this is an externalptr}

\item{type}{a string character that identifies the underlying type of this object.}
}
\value{
an ExternalObjRef if the input is an externalptr, or 'obj' otherwise.
}
\description{
Create a wrapper around an externalptr
}
