% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00class-definitions.r
\docType{class}
\name{RegularTimeSeries-class}
\alias{RegularTimeSeries-class}
\title{Class RegularTimeSeries.}
\description{
Class \code{RegularTimeSeries} is a simplified structure for multivariate regular series, readier for interop with a C API.
}
