% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00class-definitions.r
\docType{class}
\name{RegularTimeSeriesGeometry-class}
\alias{RegularTimeSeriesGeometry-class}
\title{Class RegularTimeSeriesGeometry.}
\description{
Class \code{RegularTimeSeriesGeometry} defines a class defining the geometry of time series with a time step identical for every step.
}

