% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping.r
\name{strExternalObjRef}
\alias{strExternalObjRef}
\title{A suitable method implementation for 'str', for cinterop ExternalObjRef objects}
\usage{
strExternalObjRef(x, ...)
}
\arguments{
\item{x}{an object expected to be an S4 'ExternalObjRef'}

\item{...}{potential further arguments (required for Method/Generic reasons).}
}
\description{
A suitable method implementation for 'str', for cinterop ExternalObjRef objects
}

