% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping.r
\name{updatePathWindows}
\alias{updatePathWindows}
\title{Update the PATH env var on windows, to locate an appropriate DLL dependency.}
\usage{
updatePathWindows(envVarName = "LIBRARY_PATH", libfilename = "mylib.dll")
}
\arguments{
\item{envVarName}{the environment variable to look for a root for architecture dependent libraries.}

\item{libfilename}{name of the DLL sought and that should be present in the architecture (32/64) subfolder.}
}
\value{
a character vector, the startup message string resulting from the update process.
}
\description{
Update the PATH env var on windows. This is a function meant to be used by packages' '.onLoad' functions. 
Looks in another env var, then depending on whether the current process is 32 or 64 bits build a path and looks for a specified DLL.
}
