% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cinterop-package.r
\docType{package}
\name{cinterop-package}
\alias{cinterop-package}
\alias{cinterop}
\title{Reusable constructs for interop with opaque pointers in a C API}
\description{
Reusable constructs for interop with opaque pointers in a C API
}
\details{
\tabular{ll}{
Package: \tab cinterop \cr
Type: \tab Package\cr
Version: \tab 0.2-4 \cr
Date: \tab 2017-08-09 \cr
Release Notes: \tab Add unit tests and additional time series conversions for dependent packages \cr
License: \tab LGPL-3 \cr
}

\tabular{lll}{
Version \tab Date \tab Notes \cr
}
}
\author{
Jean-Michel Perraud \email{jean-michel.perraud_at_csiro.au}
}
\keyword{C,}
\keyword{external}
\keyword{pointers}
